works but typescript unhappy
