+++ apps/shelter-web/src/app/shared/components/map/map.tsx
@@ -10,13 +10,8 @@ import {
   useApiLoadingStatus,
   useMap,
 } from '@vis.gl/react-google-maps';
-import { useAtom } from 'jotai';
 import { useCallback, useEffect, useState } from 'react';
-import { modalAtom } from '../../atoms/modalAtom';
-import { sheltersAtom } from '../../atoms/sheltersAtom';
-import { ModalAnimationEnum } from '../../modal/modal';
 import { mergeCss } from '../../utils/styles/mergeCss';
-import { ShelterCard, TShelter } from '../shelter/shelterCard';
 import {
   DEFAULT_GESTURE_HANDLING,
   DEFAULT_MAP_ZOOM,
@@ -54,8 +49,6 @@ export function Map(props: TMap) {
 
   const map = useMap();
   const mapApiStatus = useApiLoadingStatus();
-  const [_modal, setModal] = useAtom(modalAtom);
-  const [shelters] = useAtom(sheltersAtom);
 
   const [cameraProps, setCameraProps] = useState<MapCameraProps>({
     center: toGoogleLatLng(defaultCenter) as google.maps.LatLngLiteral,
@@ -65,24 +58,6 @@ export function Map(props: TMap) {
   useEffect(() => {
     console.info(`[map] loading status: ${mapApiStatus}`);
   }, [mapApiStatus]);
-
-  const handleClick = (markerId: string | null | undefined) => {
-    if (!markerId) {
-      return;
-    }
-    setModal({
-      content: (
-        <ShelterCard
-          className="mt-4"
-          shelter={
-            shelters.find((shelter) => shelter.id === markerId) as TShelter
-          }
-        />
-      ),
-      animation: ModalAnimationEnum.EXPAND,
-      closeOnMaskClick: true,
-    });
-  };
 
   const handleCameraChange = useCallback(
     (event: MapCameraChangedEvent) => {
@@ -138,7 +113,7 @@ export function Map(props: TMap) {
           key={marker.id}
           position={toGoogleLatLng(marker.position)}
           zIndex={99}
-          onClick={() => handleClick(marker.id)}
+          onClick={marker.onClick}
         >
           <MapPinIcon className="h-10" type="secondary" />
         </AdvancedMarker>
